pyproject.toml
[tool.poetry]
name = "langgraph-project"
version = "0.1.0"
description = ""
authors = ["Eden Marco <emarco177@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
beautifulsoup4 = "*"
langchain-community = "*"
tiktoken = "*"
langchainhub = "*"
langchain = "*"
langgraph = "*"
tavily-python = "*"
langchain-openai = "*"
python-dotenv = "*"
black = "*"
isort = "*"
pytest = "*"
langchain-chroma = "*"
langchain-tavily = "^0.1.5"
langchain-ollama = "^0.3.7"
langchain-google-genai = "^2.1.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"